requirements.txt
numpy>=1.19.2
opencv-python>=4.5.1
scikit-learn>=0.24.1
pillow>=8.2.0
pandas>=1.2.4
joblib>=1.0.1
matplotlib>=3.4.2
seaborn>=0.11.1
